--- apply_ffi.lua
@@ -1,203 +1,0 @@
-local ffi = require 'ffi'
-
-ffi.cdef[[
-bool THCudaTensor_pointwiseApply1(THCState* state,
-                                  THCudaTensor* a,
-                                  const char* apply_header,
-                                  const char* op_string);
-
-bool THCudaTensor_pointwiseApply2(THCState* state,
-                                  THCudaTensor* a,
-                                  THCudaTensor* b,
-                                  const char* apply_header,
-                                  const char* op_string);
-
-bool THCudaTensor_pointwiseApply3(THCState* state,
-                                  THCudaTensor* a,
-                                  THCudaTensor* b,
-                                  THCudaTensor* c,
-                                  const char* apply_header,
-                                  const char* op_string);
-]]
-
-CU.APPLY_C = ffi.load './build/libcutorchrtc.so'
-
--- copy paste from THC, could be moved to .cu
--- stays here because there is no need to put \n\ in the end
--- each line
-CU.APPLY_INCLUDE = [[
-// Maximum number of dimensions allowed for cutorch
-#define MAX_CUTORCH_DIMS 25
-
-
-// CUDA kernel argument that defines tensor layout
-template <typename IndexType>
-struct TensorInfo {
-  // Contiguous tensors of more than one dimension are collapsed down
-  // to one tensor
-  __device__ inline bool isContiguous() const {
-    return (dims == 1 && strides[0] == 1);
-  }
-
-  float* data;
-  IndexType sizes[MAX_CUTORCH_DIMS];
-  IndexType strides[MAX_CUTORCH_DIMS];
-  int dims;
-};
-
-// Translate a linear index for the apply to a float* offset;
-// specialized on `Dims` to reduce nvcc compilation time
-template <typename IndexType, int Dims>
-struct IndexToOffset {
-  static __host__ __device__ IndexType get(
-    IndexType linearId,
-    const TensorInfo<IndexType>& info) {
-    IndexType offset = 0;
-
-    // Use static dims
-    for (int i = Dims - 1; i >= 0; --i) {
-      IndexType curDimIndex = linearId % info.sizes[i];
-      IndexType curDimOffset = curDimIndex * info.strides[i];
-      offset += curDimOffset;
-
-      if (i > 0) {
-        linearId /= info.sizes[i];
-      }
-    }
-
-    return offset;
-  }
-};
-
-template <typename IndexType>
-struct IndexToOffset<IndexType, -2> {
-  static __forceinline__ __host__ __device__ IndexType
-    get(IndexType linearId, const TensorInfo<IndexType>& info) {
-    return linearId;
-  }
-};
-
-template <typename IndexType>
-struct IndexToOffset<IndexType, -1> {
-  static __forceinline__ __host__ __device__ IndexType
-    get(IndexType linearId, const TensorInfo<IndexType>& info) {
-    IndexType offset = 0;
-
-    // Use dynamic dims
-    for (int i = info.dims - 1; i >= 0; --i) {
-      IndexType curDimIndex = linearId % info.sizes[i];
-      IndexType curDimOffset = curDimIndex * info.strides[i];
-      offset += curDimOffset;
-
-      linearId /= info.sizes[i];
-    }
-
-    return offset;
-  }
-};
-
-template <typename IndexType>
-__device__ __forceinline__ IndexType getLinearBlockId() {
-  return blockIdx.z * gridDim.y * gridDim.x +
-    blockIdx.y * gridDim.x +
-    blockIdx.x;
-}
-
-// Enum that indicates whether tensor arguments are read/write or
-// read-only
-enum TensorArgType { ReadWrite, ReadOnly };
-
-// Copy operator for the pointwise apply kernel
-template <typename T>
-struct CopyOp {
-  __device__ __forceinline__ void operator()(T* dst, T* src) {
-#if __CUDA_ARCH__ >= 350
-    *dst = __ldg(src);
-#else
-    *dst = *src;
-#endif
-  }
-};
-
-//
-// This file contains pointwise operation functions and kernels that
-// work on both contiguous and non-contiguous tensor arguments of
-// arbitrary (up to MAX_CUTORCH_DIMS) dimensioned arguments without
-// copying or temporary storage.
-//
-
-// Threads per block for our apply kernel
-#define THC_APPLY_THREADS_PER_BLOCK 32 * 16
-
-template <typename Op, typename IndexType, int ADims>
-#if __CUDA_ARCH__ >= 350
-__launch_bounds__(32 * 16, 4)
-#endif
-__global__ void
-THCudaTensor_pointwiseApply1(TensorInfo<IndexType> a,
-                             IndexType totalElements,
-                             Op op) {
-  for (IndexType linearIndex = blockIdx.x * blockDim.x + threadIdx.x;
-       linearIndex < totalElements;
-       linearIndex += gridDim.x * blockDim.x) {
-    // Convert `linearIndex` into an offset of `a`
-    const IndexType aOffset =
-      IndexToOffset<IndexType, ADims>::get(linearIndex, a);
-
-    op(&a.data[aOffset]);
-  }
-}
-
-template <typename Op, typename IndexType, int ADims, int BDims>
-#if __CUDA_ARCH__ >= 350
-__launch_bounds__(32 * 16, 4)
-#endif
-__global__ void
-THCudaTensor_pointwiseApply2(TensorInfo<IndexType> a,
-                             TensorInfo<IndexType> b,
-                             IndexType totalElements,
-                             Op op) {
-  for (IndexType linearIndex = blockIdx.x * blockDim.x + threadIdx.x;
-       linearIndex < totalElements;
-       linearIndex += gridDim.x * blockDim.x) {
-    // Convert `linearIndex` into an offset of `a`
-    const IndexType aOffset =
-      IndexToOffset<IndexType, ADims>::get(linearIndex, a);
-
-    // Convert `linearIndex` into an offset of `b`
-    const IndexType bOffset =
-      IndexToOffset<IndexType, BDims>::get(linearIndex, b);
-
-    op(&a.data[aOffset], &b.data[bOffset]);
-  }
-}
-
-template <typename Op, typename IndexType, int ADims, int BDims, int CDims>
-#if __CUDA_ARCH__ >= 350
-__launch_bounds__(32 * 16, 4)
-#endif
-__global__ void
-THCudaTensor_pointwiseApply3(TensorInfo<IndexType> a,
-                             TensorInfo<IndexType> b,
-                             TensorInfo<IndexType> c,
-                             IndexType totalElements,
-                             Op op) {
-  for (IndexType linearIndex = blockIdx.x * blockDim.x + threadIdx.x;
-       linearIndex < totalElements;
-       linearIndex += gridDim.x * blockDim.x) {
-    // Convert `linearIndex` into an offset of `a`
-    const IndexType aOffset =
-      IndexToOffset<IndexType, ADims>::get(linearIndex, a);
-
-    // Convert `linearIndex` into an offset of `b`
-    const IndexType bOffset =
-      IndexToOffset<IndexType, BDims>::get(linearIndex, b);
-
-    // Convert `linearIndex` into an offset of `c`
-    const IndexType cOffset =
-      IndexToOffset<IndexType, CDims>::get(linearIndex, c);
-
-    op(&a.data[aOffset], &b.data[bOffset], &c.data[cOffset]);
-  }
-}
-]]
